feat(Input): focus on input after creating a task
Putting myself in the shoes of the user, I would want to add a few tasks before I decide to edit a task.

When I come on a todo list with the intention to add tasks, I'm likely adding a few tasks at a time, it's nice to be able to just crank them out.